Implemented session

+++ ITPE3200_Symptomizer_v2/Controllers/PatientController.cs
@@ -1,8 +1,10 @@
 ﻿using ITPE3200_Symptomizer.DAL;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
+using System.Net;
 using System.Threading.Tasks;
 using ITPE3200_Symptomizer.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 
 namespace ITPE3200_Symptomizer.Controllers
@@ -12,6 +14,7 @@ namespace ITPE3200_Symptomizer.Controllers
     {
         private readonly IPatientRepository _db;
         private ILogger<PatientController> _log;
+        private const string _loggedIn = "loggedIn";
 
         public PatientController(IPatientRepository db, ILogger<PatientController> log)
         {
@@ -20,6 +23,10 @@ namespace ITPE3200_Symptomizer.Controllers
         }
         public async Task<ActionResult> AddPatient(Patient p)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedIn)))
+            {
+                return Unauthorized();
+            }
             if (ModelState.IsValid)
             {
                 bool returnOk = await _db.AddPatient(p);
@@ -37,6 +44,10 @@ namespace ITPE3200_Symptomizer.Controllers
 
         public async Task<ActionResult> FindAll()
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedIn)))
+            {
+                return Unauthorized();
+            }
             List<Patient> allPatients = await _db.FindAll(); 
             //We can also use "var allPatients" instead of List<> (List<> usage not required)
             return Ok(allPatients);
@@ -44,6 +55,10 @@ namespace ITPE3200_Symptomizer.Controllers
 
         public async Task<ActionResult> DeletePatient(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedIn)))
+            {
+                return Unauthorized();
+            }
             bool returnOk = await _db.DeletePatient(id);
             if (!returnOk)
             {
@@ -55,6 +70,10 @@ namespace ITPE3200_Symptomizer.Controllers
 
         public async Task<ActionResult> FindPatient(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedIn)))
+            {
+                return Unauthorized();
+            }
             Patient foundPatient = await _db.FindPatient(id);
             if(foundPatient == null)
             {
@@ -65,6 +84,10 @@ namespace ITPE3200_Symptomizer.Controllers
         }
         public async Task<ActionResult> EditPatient(Patient eP)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedIn)))
+            {
+                return Unauthorized();
+            }
             if (ModelState.IsValid)
             {
                 bool returnOk = await _db.EditPatient(eP);
@@ -87,12 +110,18 @@ namespace ITPE3200_Symptomizer.Controllers
                 if (!returnOk)
                 {
                     _log.LogInformation("Failed to logg in with user name {user.Username}", user.Username);
+                    HttpContext.Session.SetString(_loggedIn,"");
                     return Ok(false);
                 }
+                HttpContext.Session.SetString(_loggedIn, "loggedIn");
                 return Ok(true);
             }
             _log.LogInformation("Fail in input validation");
             return BadRequest("Fail in input validation on server side");
+        }
+        public void LoggOut()
+        {
+            HttpContext.Session.SetString(_loggedIn,"");
         }
     }
 }
